Address intermittent shakedown problems
Problems fixed:
- [SPARK] Teardown spark-standalone when finished. Otherwise if hdfs runs after it on a 5-node cluster, it can get stranded trying to find a place to run the name-1-bootstrap task. Other frameworks likely run into similar issues when they come after spark-standalone.
- [HDFS] Increased mem on data/name/journal/zk nodes from 512M to 2048M. Default heap is 1000M, so container should be 2x that based on https://github.com/mesosphere/dcos-commons/blob/0fd04114049e09ca9e975c79a42f0652f803a8a3/frameworks/hdfs/docs/configure.md#L428-L428
- [ELASTIC] Fixed logic error in test_kibana_proxylite_adminrouter_integration that should have failed 100% of the time, but instead only failed once in ~20 runs.

Ancillary changes:
- Disable teardown of cluster so Mesos logs are around for post-failure inspection. Clusters last for 4 hours and tests take 3 to run. I think it's worth overpaying a few dollars for each CI run until the test suite stabilizes.
- Combined some HDFS shakedown files to shave 20-30 minutes off CI runs
- Removed unused test_upgrade.py in HDFS (can always add it back in later when it's ready to be used)
- Created marathon config update helper method with assertion of success to guard against future regressions
- Fixed some unused imports, variables, and PEP-8 formatting complaints
- Added some debug information around HDFS package installation, as that's where a lot of failures seem to be happening
- Simplified lambda in sdk_marathon and made api_url usage more consistent
- Removed unused marathon_api_url method from elastic config.py
