+++ frameworks/spark/tests/test_sanity.py
@@ -1,6 +1,7 @@
 import pytest
 
 import sdk_install as install
+import sdk_utils as utils
 
 from tests.config import (
     PACKAGE_NAME,
@@ -10,10 +11,15 @@ from tests.config import (
 
 def setup_module(module):
     install.uninstall(PACKAGE_NAME)
+    utils.gc_frameworks()
     install.install(PACKAGE_NAME, DEFAULT_TASK_COUNT)
+
+
+def teardown_module(module):
+    install.uninstall(PACKAGE_NAME)
 
 
 @pytest.mark.sanity
 @pytest.mark.smoke
 def test_install():
-    pass # package installed and appeared healthy!
+    pass  # package installed and appeared healthy!
