+++ testing/sdk_plan.py
@@ -1,7 +1,6 @@
 '''Utilities relating to interaction with service plans'''
 
 import dcos
-import sdk_cmd
 import sdk_spin
 import shakedown
 
@@ -19,16 +18,14 @@ def start_sidecar_plan(service_name, parameters=None):
 
 
 def start_plan(service_name, plan, parameters=None):
-    return dcos.http.post(
-        shakedown.dcos_service_url(service_name) +
-            "/v1/plans/{}/start".format(plan),
-        json=parameters if parameters is not None else {})
+    return dcos.http.post("{}/v1/plans/{}/start".format(shakedown.dcos_service_url(service_name), plan),
+                          json=parameters if parameters is not None else {})
 
 
 def get_plan(service_name, plan):
     def fn():
-        response = dcos.http.get("{}/v1/plans/{}".format(
-            shakedown.dcos_service_url(service_name), plan))
+        response = dcos.http.get("{}/v1/plans/{}".format(shakedown.dcos_service_url(service_name), plan))
         response.raise_for_status()
         return response
+
     return sdk_spin.time_wait_return(lambda: fn())
